FinanceApp/stockProject - main/loadTestingApp/requirements.txt
psycopg2-binary
gevent>=1.0.0
faker
locust